pyproject.toml
[tool.poetry]
name = "extinct"
version = "0.1.0"
description = ""
authors = ["Your Name <you@example.com>"]
license = "Apache 2"

[tool.poetry.dependencies]
fair-bolts = {git = "https://github.com/predictive-analytics-lab/fair-bolts.git", branch = "main"}
hydra-submitit-launcher = {version = "^1.1.1", allow-prereleases = true}
neoconfigen = {git = "https://github.com/predictive-analytics-lab/neoconfigen.git", branch = "master"}
palkit = "*"
python = "^3.9"
pytorch-lightning = "*"
wandb = "*"
albumentations = "^1.0.0"

[tool.poetry.dev-dependencies]
black = "*"
isort = "*"
mypy = "*"
pre-commit = "*"
pydocstyle = "*"
pytest = "*"
python-type-stubs = {git = "https://github.com/predictive-analytics-lab/python-type-stubs.git", branch = "main"}
pylint = "^2.8.2"

[tool.poetry.scripts]
rawr = "extinct.main:launcher"

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"

[tool.black]
line-length = 100
target-version = ['py38']
skip-string-normalization = true
include = '\.pyi?$'
exclude = '''
(
  /(
      \.eggs         # exclude a few common directories in the
    | \.git          # root of the project
    | \.hg
    | \.mypy_cache
    | \.tox
    | \.venv
    | _build
    | buck-out
    | build
    | dist
  )/
  | foo.py           # also separately exclude a file named foo.py in
                     # the root of the project
)
'''

[tool.isort]
known_third_party = ["PIL", "albumentations", "ethicml", "fair_bolts", "hydra", "kit", "numpy", "omegaconf", "pandas", "pytest", "pytorch_lightning", "torch", "torchmetrics", "tqdm", "typing_extensions"]
known_future_library = []
extra_standard_library = ["__future__", "typing_extensions"]
line_length = 88
profile = "black"
force_sort_within_sections = "True"
classes = ["MISSING"]
